move deduplication into a test and rely upon controller framework to do this normally
