+++ internal/mesh/internal/controllers/routes/util.go
@@ -5,9 +5,6 @@ package routes
 
 import (
 	"google.golang.org/protobuf/proto"
-
-	"github.com/hashicorp/consul/internal/controller"
-	"github.com/hashicorp/consul/internal/mesh/internal/controllers/routes/xroutemapper"
 )
 
 func protoClone[T proto.Message](v T) T {
@@ -23,9 +20,4 @@ func protoSliceClone[T proto.Message](in []T) []T {
 		out = append(out, protoClone[T](v))
 	}
 	return out
-}
-
-// Deprecated: xroutemapper.DeduplicateRequests
-func deduplicate(reqs []controller.Request) []controller.Request {
-	return xroutemapper.DeduplicateRequests(reqs)
 }
